Reduce the user's distress if Rivet's API introduces a backwards-incompatible change.
Currently, Rivet executes every migration file even if that file won't
be pushed or poppped. This makes it hard for users to upgrade Rivet,
especially if they have a lot of migrations.

Instead, Rivet should only execute a migration if, when pushing, a
migration hasn't been applied to a database, and, when poppsing, a
migration has been applied to a database.