Refactoring.
When investigating why a non-migration object was somehow getting returned to Update-Database, realized that Rivet's
internals are starting to leak abstractions, so I've tidied things up.

Get-MigrationFile no longer needs a Rivet session. Its job is to return migration file info objects and parse their
file names into IDs and names. Replaced the session parameter with a database name parameter and updated the path
parameter to take a single path. Get-MigrationFile now returns all the migration files for a single database.

Update-Database should be responsible for pushing/popping unapplied migrations to a database. It had a secondary,
seperate responsibility of applying Rivet's internal migrations to a database. Merged this second responsibility into
the first so that now Update-Database knows about and applies Rivet's internal migrations to a database if they haven't
been applied. Introduced the concept to migration files of whether or not they are a Rivet internal migration so that
Update-Database would only ever push those migrations and never pop them.

Why does Invoke-Rivet import Rivet plug-ins? It should just be determining what Rivet commands to run and not worry
about setting anything up for those commands (beyond a session) or initializing the Rivet schema in a database.
Managing the plug-ins is now part of Convert-FileInfoToMigration, which has the responsibility of converting migration
scripts into operation objects. Since plug-ins interact with that process, that's where the plug-ins should be
imported. Invoke-Rivet also cared and knew about the paths to migration directories, which it shouldn't.

The Get-Migration function had some of the same problems as Invoke-Rivet: it was managing things it shouldn't be
responsible for, like loading plug-ins. Other refactorings also impacted how this function worked.